ContextCraftPro/tests/test_config.py
"""
Tests for configuration management
"""

import pytest


class TestConfig:
    """Test suite for configuration functionality"""

    def test_load_config(self, sample_config, tmp_path):
        """Test loading configuration from YAML"""
        # TODO: Implement when ccp_config is ready
        pass

    def test_config_validation(self):
        """Test configuration validation"""
        # TODO: Implement when ccp_config is ready
        pass

    def test_environment_variable_override(self, monkeypatch):
        """Test that environment variables override config values"""
        # TODO: Implement when ccp_config is ready
        pass

    def test_default_config_creation(self, tmp_path):
        """Test creation of default config file"""
        # TODO: Implement when ccp_config is ready
        pass
